+++ src/components/BurgerConstructor/styles.module.css
@@ -6,6 +6,7 @@
 .list {
     display: flex;
     flex-direction: column;
+    align-items: end;
     gap: 16px;
     max-height: calc(100vh - 344px);
     overflow-y: auto;
@@ -23,4 +24,19 @@
     height: 36px;
     width: 36px;
     margin: 0 40px 0 8px;
+}
+
+.list_el {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+    width: 100%;
+}
+
+.spacer {
+    width: 24px;
+}
+
+.dragIcon {
+    cursor: pointer;
 }
