+++ src/components/Modal/components/index.ts
@@ -1,0 +1,1 @@
+export { ModalOverlay } from "./ModalOverlay.tsx";
